Updating log limit logic
