+++ rskj-core/src/main/resources/expected.conf
@@ -282,10 +282,10 @@ rpc = {
     skipRemasc: <enabled>
     zeroSignatureIfRemasc = <enabled>
     gasEstimationCap = <gas>
-    eth_getLogs =  {
-            maxBlocksToQuery = <number>
-            maxLogsToReturn = <number>
-        }
+    logs = {
+        maxBlocksToQuery = <number>
+        maxLogsToReturn = <number>
+    }
 }
 wire = {
     protocol = <protocol>
